Using smart/dumb component pattern. Using geoserve place-list with bound _event getter/setter pattern. No subscriptions.
